+++ Libplanet/Blocks/Block.cs
@@ -25,11 +25,14 @@ namespace Libplanet.Blocks
         /// </summary>
         public const int CurrentProtocolVersion = BlockHeader.CurrentProtocolVersion;
 
-        private int _bytesLength;
+        private int? _bytesLength = null;
+        private BlockHeader? _header = null;
+        private ImmutableArray<byte>? _preEvaluationHash = null;
+        private BlockHash? _hash = null;
 
         /// <summary>
-        /// Creates a <see cref="Block{T}"/> instance by manually filling all field values.
-        /// For a more automated way, see also <see cref="Mine"/> method.
+        /// Creates a <see cref="Block{T}"/> instance by manually filling field values.
+        /// For a more automated way, see also the <see cref="Mine"/> method.
         /// </summary>
         /// <param name="index">The height of the block to create.  Goes to the <see cref="Index"/>.
         /// </param>
@@ -60,12 +63,17 @@ namespace Libplanet.Blocks
         /// </param>
         /// <param name="protocolVersion">The protocol version. <see cref="CurrentProtocolVersion"/>
         /// by default.</param>
-        /// <exception cref="ArgumentNullException">Thrown when both
-        /// <paramref name="hashAlgorithm"/> and <paramref name="preEvaluationHash"/> are omitted.
+        /// <exception cref="ArgumentException">Thrown when either both
+        /// <paramref name="hashAlgorithm"/> and <paramref name="preEvaluationHash"/> are present
+        /// or both are missing.
+        /// <exception cref="ArgumentException">Thrown when <paramref name="stateRootHash"/>
+        /// is null while <paramref name="preEvaluationHash"/> is not null.
         /// </exception>
-        /// <exception cref="ArgumentException">Thrown when both <paramref name="hashAlgorithm"/>
-        /// and <paramref name="preEvaluationHash"/>, which are mutually exclusive, are present.
         /// </exception>
+        /// <remarks>
+        /// Due to historic reasons, there is non-trivial implicit logic embedded inside this
+        /// constructor.  It is strongly recommended to use <see cref="Mine"/> instead.
+        /// </remarks>
         /// <seealso cref="Mine"/>
         public Block(
             long index,
@@ -81,6 +89,28 @@ namespace Libplanet.Blocks
             HashDigest<SHA256>? stateRootHash = null,
             int protocolVersion = CurrentProtocolVersion)
         {
+            // FIXME: This constructor needs to be separated into several overloads.
+            // Let (bool h, bool p, bool s) represent non-null arguments provided for
+            // hashAlgorithm, preEvaluationHash, and stateRootHash respectively.
+            // Despite its summary, currently accepted only use cases are:
+            //  - (true, false, false): Called when preEvaluationHash needs to be calculated.
+            //  - (false, true, false): Allowed purely for testing.  Normal code path should not
+            //    lead to this combination.
+            //  - (false, true, true): Called when stateRootHash needs to be attached.
+            // All other combinations are rejected.
+            if (!(hashAlgorithm is { } ^ preEvaluationHash is { }))
+            {
+                throw new ArgumentException(
+                    $"Exactly one of {nameof(hashAlgorithm)} " +
+                    $"and {nameof(preEvaluationHash)} should be provided as non-null.");
+            }
+            else if (!(preEvaluationHash is { }) && stateRootHash is { })
+            {
+                throw new ArgumentException(
+                    $"Parameter {nameof(stateRootHash)} cannot be non-null while" +
+                    $"{nameof(preEvaluationHash)} is null.");
+            }
+
             ProtocolVersion = protocolVersion;
             Index = index;
             Difficulty = difficulty;
@@ -92,37 +122,58 @@ namespace Libplanet.Blocks
             Transactions = transactions.OrderBy(tx => tx.Id).ToArray();
             TxHash = CalculateTxHashes(Transactions);
 
-            // FIXME: This constructor needs to be separated into two overloads, which are one
-            // taking only hashAlgorithm and other one taking only preEvaluationHash, so that
-            // the compiler can check whether two parameters are both omitted or both passed
-            // if any chance.
-            if (preEvaluationHash is { } hash)
+#pragma warning disable SA1118
+            if (hashAlgorithm is { } ha)
             {
-                if (hashAlgorithm is { })
-                {
-                    throw new ArgumentException(
-                        $"The parameters {nameof(hashAlgorithm)} and {nameof(preEvaluationHash)} " +
-                        "are mutually exclusive.",
-                        paramName: nameof(preEvaluationHash)
-                    );
-                }
-
-                PreEvaluationHash = hash;
-            }
-            else if (hashAlgorithm is { } algo)
-            {
-                PreEvaluationHash = algo.Digest(Header.SerializeForHash()).ToImmutableArray();
+                // FIXME: This only works due to sanity constraint on usage.
+                _header = new BlockHeader(
+                    protocolVersion: ProtocolVersion,
+                    index: Index,
+                    timestamp: Timestamp.ToString(
+                        BlockHeader.TimestampFormat,
+                        CultureInfo.InvariantCulture),
+                    nonce: Nonce.ToByteArray().ToImmutableArray(),
+                    miner: Miner.ToByteArray().ToImmutableArray(),
+                    difficulty: Difficulty,
+                    totalDifficulty: TotalDifficulty,
+                    previousHash: PreviousHash?.ToByteArray().ToImmutableArray()
+                        ?? ImmutableArray<byte>.Empty,
+                    txHash: TxHash?.ToByteArray().ToImmutableArray()
+                        ?? ImmutableArray<byte>.Empty,
+                    hashAlgorithm: ha);
+                _preEvaluationHash = Header.PreEvaluationHash;
+                StateRootHash = stateRootHash;
+                _hash = new BlockHash(Header.Hash);
             }
             else
             {
-                string message = $"Either {nameof(hashAlgorithm)} or {nameof(preEvaluationHash)} " +
-                    "must to be passed.";
-                throw new ArgumentNullException(nameof(preEvaluationHash), message);
+                ImmutableArray<byte> peh = preEvaluationHash
+                    ?? throw new NullReferenceException(
+                        $"Parameter {nameof(preEvaluationHash)} cannot be null.");
+
+                _header = new BlockHeader(
+                    protocolVersion: ProtocolVersion,
+                    index: Index,
+                    timestamp: Timestamp.ToString(
+                        BlockHeader.TimestampFormat,
+                        CultureInfo.InvariantCulture),
+                    nonce: Nonce.ToByteArray().ToImmutableArray(),
+                    miner: Miner.ToByteArray().ToImmutableArray(),
+                    difficulty: Difficulty,
+                    totalDifficulty: TotalDifficulty,
+                    previousHash: PreviousHash?.ToByteArray().ToImmutableArray()
+                        ?? ImmutableArray<byte>.Empty,
+                    txHash: TxHash?.ToByteArray().ToImmutableArray()
+                        ?? ImmutableArray<byte>.Empty,
+                    preEvaluationHash: peh,
+                    stateRootHash: stateRootHash?.ToByteArray().ToImmutableArray()
+                        ?? ImmutableArray<byte>.Empty);
+
+                _preEvaluationHash = Header.PreEvaluationHash;
+                StateRootHash = stateRootHash;
+                _hash = new BlockHash(Header.Hash);
             }
-
-            StateRootHash = stateRootHash;
-
-            Hash = BlockHash.DeriveFrom(Header.SerializeForHash());
+#pragma warning restore SA1118
 
             // As the order of transactions should be unpredictable until a block is mined,
             // the sorter key should be derived from both a block hash and a txid.
@@ -170,12 +221,20 @@ namespace Libplanet.Blocks
         {
         }
 
-        // FIXME: Should this necessarily be a public constructor?
+#pragma warning disable SA1514, SA1515
+        // FIXME: This probably should not be public.
         // See also <https://github.com/planetarium/libplanet/issues/1146>.
+        /// <summary>
+        /// Creates a <see cref="Block{T}"/> instance by attaching <paramref name="stateRootHash"/>
+        /// to given <paramref name="block"/>.
+        /// </summary>
+        /// <param name="block">A <see cref="Block{T}"/> instance without
+        /// <see cref="Block{T}.StateRootHash"/> to use.</param>
+        /// <param name="stateRootHash">A <see cref="HashDigest{SHA256}"/> to attach.</param>
+#pragma warning restore SA1514, SA1515
         public Block(
             Block<T> block,
-            HashDigest<SHA256>? stateRootHash
-        )
+            HashDigest<SHA256> stateRootHash)
             : this(
                 block.Index,
                 block.Difficulty,
@@ -188,78 +247,51 @@ namespace Libplanet.Blocks
                 hashAlgorithm: null,
                 preEvaluationHash: block.PreEvaluationHash,
                 stateRootHash: stateRootHash,
-                protocolVersion: block.ProtocolVersion
-            )
+                protocolVersion: block.ProtocolVersion)
         {
         }
 
-        private Block(RawBlock rb)
-            : this(
-#pragma warning disable SA1118
-                rb.Header.ProtocolVersion,
-                new BlockHash(rb.Header.Hash),
-                rb.Header.Index,
-                rb.Header.Difficulty,
-                rb.Header.TotalDifficulty,
-                new Nonce(rb.Header.Nonce.ToArray()),
-                new Address(rb.Header.Miner),
-                rb.Header.PreviousHash.Any()
-                    ? new BlockHash(rb.Header.PreviousHash)
-                    : (BlockHash?)null,
-                DateTimeOffset.ParseExact(
-                    rb.Header.Timestamp,
-                    BlockHeader.TimestampFormat,
-                    CultureInfo.InvariantCulture).ToUniversalTime(),
-                rb.Header.TxHash.Any()
-                    ? new HashDigest<SHA256>(rb.Header.TxHash)
-                    : (HashDigest<SHA256>?)null,
-                rb.Transactions
-                    .Select(tx => Transaction<T>.Deserialize(tx.ToArray(), false))
-                    .ToList(),
-                rb.Header.PreEvaluationHash.Any()
-                    ? rb.Header.PreEvaluationHash
-                    : (ImmutableArray<byte>?)null,
-                rb.Header.StateRootHash.Any()
-                    ? new HashDigest<SHA256>(rb.Header.StateRootHash)
-                    : (HashDigest<SHA256>?)null)
-#pragma warning restore SA1118
+        private Block(RawBlock rawBlock)
         {
-        }
+            _header = rawBlock.Header;
 
-        private Block(
-            int protocolVersion,
-            BlockHash hash,
-            long index,
-            long difficulty,
-            BigInteger totalDifficulty,
-            Nonce nonce,
-            Address miner,
-            BlockHash? previousHash,
-            DateTimeOffset timestamp,
-            HashDigest<SHA256>? txHash,
-            IReadOnlyList<Transaction<T>> transactions,
-            ImmutableArray<byte>? preEvaluationHash,
-            HashDigest<SHA256>? stateRootHash
-        )
-        {
-            ProtocolVersion = protocolVersion;
-            Index = index;
-            Difficulty = difficulty;
-            TotalDifficulty = totalDifficulty;
-            Nonce = nonce;
-            Miner = miner;
-            PreviousHash = previousHash;
-            Timestamp = timestamp;
-            Hash = hash;
-            PreEvaluationHash = preEvaluationHash ??
-                throw new ArgumentNullException(nameof(preEvaluationHash));
+            ProtocolVersion = rawBlock.Header.ProtocolVersion;
+            Index = rawBlock.Header.Index;
+            Difficulty = rawBlock.Header.Difficulty;
+            TotalDifficulty = rawBlock.Header.TotalDifficulty;
+            Nonce = new Nonce(rawBlock.Header.Nonce.ToArray());
+            Miner = new Address(rawBlock.Header.Miner);
+            PreviousHash = rawBlock.Header.PreviousHash.Any()
+                ? new BlockHash(rawBlock.Header.PreviousHash)
+                : (BlockHash?)null;
+            Timestamp = DateTimeOffset.ParseExact(
+                Header.Timestamp,
+                BlockHeader.TimestampFormat,
+                CultureInfo.InvariantCulture).ToUniversalTime();
+            TxHash = Header.TxHash.Any()
+                ? new HashDigest<SHA256>(rawBlock.Header.TxHash)
+                : (HashDigest<SHA256>?)null;
 
-            // See also: https://github.com/planetarium/libplanet/pull/1116#discussion_r535836480
-            // FIXME: we should convert `StateRootHash`'s type to `HashDisgest<SHA256>` after
+            // FIXME: Transactions should be re-ordered to properly validate StateRootHash.
+            // See also <https://github.com/planetarium/libplanet/issues/1299>.
+            Transactions = rawBlock.Transactions
+                .Select(tx => Transaction<T>.Deserialize(tx.ToArray(), false))
+                .ToImmutableList();
+
+            _preEvaluationHash = rawBlock.Header.PreEvaluationHash.Any()
+                ? rawBlock.Header.PreEvaluationHash
+                : throw new ArgumentException(
+                    $"PreEvaluationHash of {nameof(rawBlock.Header)} cannot be empty.");
+
+            // FIXME: We should convert `StateRootHash`'s type to `HashDisgest<SHA256>` after
             // removing `IBlockStateStore`.
-            StateRootHash = stateRootHash;
-            TxHash = txHash;
-            Transactions = transactions.ToImmutableArray();
+            // See also <https://github.com/planetarium/libplanet/pull/1116#discussion_r535836480>.
+            // FIXME: Normal path should not lead to StateRootHash being null.  Should be
+            // refactored to throw an exception.
+            StateRootHash = rawBlock.Header.StateRootHash.Any()
+                ? new HashDigest<SHA256>(rawBlock.Header.StateRootHash)
+                : (HashDigest<SHA256>?)null;
+            _hash = new BlockHash(rawBlock.Header.Hash);
         }
 
         /// <summary>
@@ -274,7 +306,8 @@ namespace Libplanet.Blocks
         /// </summary>
         /// <seealso cref="PreEvaluationHash"/>
         /// <seealso cref="StateRootHash"/>
-        public BlockHash Hash { get; }
+        public BlockHash Hash => _hash
+            ?? throw new InvalidOperationException("Hash has not been set.");
 
         /// <summary>
         /// The hash derived from the block <em>except of</em>
@@ -283,7 +316,8 @@ namespace Libplanet.Blocks
         /// </summary>
         /// <seealso cref="Nonce"/>
         /// <seealso cref="BlockHeader.Validate"/>
-        public ImmutableArray<byte> PreEvaluationHash { get; }
+        public ImmutableArray<byte> PreEvaluationHash => _preEvaluationHash
+            ?? throw new InvalidOperationException("PreEvaluationHash is has not been set.");
 
         /// <summary>
         /// The <see cref="ITrie.Hash"/> of the states on the block.
@@ -330,9 +364,7 @@ namespace Libplanet.Blocks
         {
             get
             {
-                // Note that Serialize() by itself caches _byteLength, so that this ByteLength
-                // property never invokes Serialize() more than once.
-                return _bytesLength > 0 ? _bytesLength : Serialize().Length;
+                return _bytesLength ?? (int)(_bytesLength = Serialize().Length);
             }
         }
 
@@ -340,36 +372,8 @@ namespace Libplanet.Blocks
         /// The <see cref="BlockHeader"/> of the block.
         /// </summary>
         [IgnoreDuringEquals]
-        public BlockHeader Header
-        {
-            get
-            {
-                string timestampAsString = Timestamp.ToString(
-                    BlockHeader.TimestampFormat,
-                    CultureInfo.InvariantCulture
-                );
-                ImmutableArray<byte> previousHashAsArray =
-                    PreviousHash?.ToByteArray().ToImmutableArray() ?? ImmutableArray<byte>.Empty;
-                ImmutableArray<byte> stateRootHashAsArray =
-                    StateRootHash?.ToByteArray().ToImmutableArray() ?? ImmutableArray<byte>.Empty;
-
-                // FIXME: When hash is not assigned, should throw an exception.
-                return new BlockHeader(
-                    protocolVersion: ProtocolVersion,
-                    index: Index,
-                    timestamp: timestampAsString,
-                    nonce: Nonce.ToByteArray().ToImmutableArray(),
-                    miner: Miner.ToByteArray().ToImmutableArray(),
-                    difficulty: Difficulty,
-                    totalDifficulty: TotalDifficulty,
-                    previousHash: previousHashAsArray,
-                    txHash: TxHash?.ToByteArray().ToImmutableArray() ?? ImmutableArray<byte>.Empty,
-                    hash: Hash.ToByteArray().ToImmutableArray(),
-                    preEvaluationHash: PreEvaluationHash,
-                    stateRootHash: stateRootHashAsArray
-                );
-            }
-        }
+        public BlockHeader Header => _header
+            ?? throw new InvalidOperationException("Header has not been set.");
 
         public static bool operator ==(Block<T> left, Block<T> right) =>
             Operator.Weave(left, right);
@@ -496,7 +500,6 @@ namespace Libplanet.Blocks
         {
             var codec = new Codec();
             byte[] serialized = codec.Encode(ToBencodex());
-            _bytesLength = serialized.Length;
             return serialized;
         }
 
@@ -553,24 +556,6 @@ namespace Libplanet.Blocks
             foreach (Transaction<T> tx in Transactions)
             {
                 tx.Validate();
-            }
-
-            if (ProtocolVersion > 0)
-            {
-                byte[] expectedPreEvaluationHash =
-                    hashAlgorithm.Digest(Header.SerializeForHash(includeStateRootHash: false));
-                if (!ByteUtil.TimingSafelyCompare(expectedPreEvaluationHash, PreEvaluationHash))
-                {
-                    string message =
-                        $"The expected pre evaluation hash of block {Hash} is " +
-                        $"{ByteUtil.Hex(expectedPreEvaluationHash)}, but its pre evaluation hash " +
-                        $"is {ByteUtil.Hex(PreEvaluationHash)}.";
-                    throw new InvalidBlockPreEvaluationHashException(
-                        PreEvaluationHash,
-                        expectedPreEvaluationHash.ToImmutableArray(),
-                        message
-                    );
-                }
             }
 
             HashDigest<SHA256>? calculatedTxHash =