+++ Libplanet/Blocks/BlockHeader.cs
@@ -49,35 +49,37 @@ namespace Libplanet.Blocks
         private static readonly Codec Codec = new Codec();
 
         /// <summary>
-        /// Creates a <see cref="BlockHeader"/> instance.
+        /// Creates a <see cref="BlockHeader"/> instance by manually filling in all the properties.
         /// </summary>
-        /// <param name="protocolVersion">The protocol version.  Goes to the <see
-        /// cref="ProtocolVersion"/>.</param>
-        /// <param name="index">The height of the block.  Goes to the <see cref="Index"/>.
+        /// <param name="protocolVersion">The protocol version.  Goes to
+        /// <see cref="ProtocolVersion"/>.</param>
+        /// <param name="index">The height of the block.  Goes to <see cref="Index"/>.
         /// </param>
-        /// <param name="timestamp">The time this block is created.
-        /// Goes to the <see cref="Timestamp"/>.</param>
-        /// <param name="nonce">The nonce which satisfy the given <paramref name="difficulty"/>
-        /// with any other field values.  Goes to the <see cref="Nonce"/>.</param>
-        /// <param name="miner">An optional address refers to who mines this block.
-        /// Goes to the <see cref="Miner"/>.</param>
+        /// <param name="timestamp">The time the block is created.
+        /// Goes to <see cref="Timestamp"/>.</param>
+        /// <param name="nonce">The nonce which satisfies given <paramref name="difficulty"/>.
+        /// Goes to <see cref="Nonce"/>.</param>
+        /// <param name="miner">The miner of the block.  Goes to <see cref="Miner"/>.</param>
         /// <param name="difficulty">The mining difficulty that <paramref name="nonce"/>
-        /// has to satisfy.  Goes to the <see cref="Difficulty"/>.</param>
-        /// <param name="totalDifficulty">The total mining difficulty until this block.
-        /// See also <see cref="Difficulty"/>.</param>
+        /// has to satisfy.  Goes to <see cref="Difficulty"/>.</param>
+        /// <param name="totalDifficulty">The total mining difficulty since the genesis,
+        /// including the block's difficulty.  See also <see cref="Difficulty"/>.</param>
         /// <param name="previousHash">The previous block's <see cref="Hash"/>.  If it's a genesis
-        /// block (i.e., <paramref name="index"/> is 0) this should be <c>null</c>.
-        /// Goes to the <see cref="PreviousHash"/>.</param>
+        /// block (i.e., its <see cref="Block{T}.Index"/> is 0) this should be empty.
+        /// Goes to <see cref="PreviousHash"/>.</param>
         /// <param name="txHash">The result of hashing the transactions the block has.
-        /// Goes to the <see cref="TxHash"/>.</param>
+        /// Goes to <see cref="TxHash"/>.</param>
         /// <param name="hash">The hash of the <see cref="Block{T}"/>.
-        /// Goes to the <see cref="Hash"/>.</param>
-        /// <param name="preEvaluationHash">The hash derived from the block <em>except of</em>
+        /// Goes to <see cref="Hash"/>.</param>
+        /// <param name="preEvaluationHash">The hash derived from the block <em>excluding</em>
         /// <paramref name="stateRootHash"/> (i.e., without action evaluation).
-        /// Used for <see cref="Validate"/> checking <paramref name="nonce"/>.
-        /// </param>
+        /// Used for checking <paramref name="nonce"/>.  See also <see cref="Validate"/>.</param>
         /// <param name="stateRootHash">The <see cref="ITrie.Hash"/> of the states on the block.
         /// </param>
+        /// <remarks>
+        /// This is only exposed for testing. Should not be used as an entry point to create
+        /// a <see cref="BlockHeader"/> instance under normal circumstances.
+        /// </remarks>
         public BlockHeader(
             int protocolVersion,
             long index,
@@ -106,6 +108,12 @@ namespace Libplanet.Blocks
             StateRootHash = stateRootHash;
         }
 
+        /// <summary>
+        /// Creates a <see cref="BlockHeader"/> instance from its serialization.
+        /// </summary>
+        /// <param name="dict">The <see cref="Bencodex.Types.Dictionary"/>
+        /// representation of <see cref="BlockHeader"/> instance.
+        /// </param>
         public BlockHeader(Bencodex.Types.Dictionary dict)
         {
             ProtocolVersion = dict.ContainsKey(ProtocolVersionKey)
@@ -126,10 +134,6 @@ namespace Libplanet.Blocks
                 ? dict.GetValue<Binary>(TxHashKey).ToImmutableArray()
                 : ImmutableArray<byte>.Empty;
 
-            Hash = dict.ContainsKey((IKey)(Binary)HashKey)
-                ? dict.GetValue<Binary>(HashKey).ToImmutableArray()
-                : ImmutableArray<byte>.Empty;
-
             PreEvaluationHash = dict.ContainsKey((IKey)(Binary)PreEvaluationHashKey)
                 ? dict.GetValue<Binary>(PreEvaluationHashKey).ToImmutableArray()
                 : ImmutableArray<byte>.Empty;
@@ -137,6 +141,118 @@ namespace Libplanet.Blocks
             StateRootHash = dict.ContainsKey((IKey)(Binary)StateRootHashKey)
                 ? dict.GetValue<Binary>(StateRootHashKey).ToImmutableArray()
                 : ImmutableArray<byte>.Empty;
+
+            Hash = dict.ContainsKey((IKey)(Binary)HashKey)
+                ? dict.GetValue<Binary>(HashKey).ToImmutableArray()
+                : ImmutableArray<byte>.Empty;
+        }
+
+        /// <summary>
+        /// Creates a <see cref="BlockHeader"/> instance for a <see cref="Block{T}"/> instance with
+        /// missing <see cref="Block{T}.StateRootHash"/>.
+        /// </summary>
+        /// <param name="protocolVersion">The protocol version.  Goes to
+        /// <see cref="ProtocolVersion"/>.</param>
+        /// <param name="index">The height of the block.  Goes to <see cref="Index"/>.
+        /// </param>
+        /// <param name="timestamp">The time the block is created.
+        /// Goes to <see cref="Timestamp"/>.</param>
+        /// <param name="nonce">The nonce which satisfies given <paramref name="difficulty"/>.
+        /// Goes to <see cref="Nonce"/>.</param>
+        /// <param name="miner">The miner of the block.  Goes to <see cref="Miner"/>.</param>
+        /// <param name="difficulty">The mining difficulty that <paramref name="nonce"/>
+        /// has to satisfy.  Goes to <see cref="Difficulty"/>.</param>
+        /// <param name="totalDifficulty">The total mining difficulty since the genesis
+        /// including the block's difficulty.  See also <see cref="Difficulty"/>.</param>
+        /// <param name="previousHash">The previous block's <see cref="Hash"/>.  If it's a genesis
+        /// block (i.e., its <see cref="Block{T}.Index"/> is 0) this should be empty.
+        /// Goes to <see cref="PreviousHash"/>.</param>
+        /// <param name="txHash">The result of hashing the transactions the block has.
+        /// Goes to <see cref="TxHash"/>.</param>
+        /// <param name="hashAlgorithm">The proof-of-work hash algorithm.</param>
+        internal BlockHeader(
+            int protocolVersion,
+            long index,
+            string timestamp,
+            ImmutableArray<byte> nonce,
+            ImmutableArray<byte> miner,
+            long difficulty,
+            BigInteger totalDifficulty,
+            ImmutableArray<byte> previousHash,
+            ImmutableArray<byte> txHash,
+            HashAlgorithmType hashAlgorithm)
+        {
+            ProtocolVersion = protocolVersion;
+            Index = index;
+            Timestamp = timestamp;
+            Nonce = nonce;
+            Miner = miner;
+            Difficulty = difficulty;
+            TotalDifficulty = totalDifficulty;
+            PreviousHash = previousHash;
+            TxHash = txHash;
+
+            byte[] serialized = SerializeForPreEvaluationHash();
+            PreEvaluationHash = hashAlgorithm.Digest(serialized).ToImmutableArray();
+            StateRootHash = ImmutableArray<byte>.Empty;
+            Hash = BlockHash.DeriveFrom(serialized).ByteArray;
+        }
+
+        /// <summary>
+        /// Creates a <see cref="BlockHeader"/> instance for a <see cref="Block{T}"/>.
+        /// </summary>
+        /// <param name="protocolVersion">The protocol version.  Goes to
+        /// <see cref="ProtocolVersion"/>.</param>
+        /// <param name="index">The height of the block.  Goes to <see cref="Index"/>.
+        /// </param>
+        /// <param name="timestamp">The time the block is created.
+        /// Goes to <see cref="Timestamp"/>.</param>
+        /// <param name="nonce">The nonce which satisfies given <paramref name="difficulty"/>.
+        /// Goes to <see cref="Nonce"/>.</param>
+        /// <param name="miner">The miner of the block.  Goes to <see cref="Miner"/>.</param>
+        /// <param name="difficulty">The mining difficulty that <paramref name="nonce"/>
+        /// has to satisfy.  Goes to <see cref="Difficulty"/>.</param>
+        /// <param name="totalDifficulty">The total mining difficulty since the genesis
+        /// including the block's difficulty.  See also <see cref="Difficulty"/>.</param>
+        /// <param name="previousHash">The previous block's <see cref="Hash"/>.  If it's a genesis
+        /// block (i.e., its <see cref="Block{T}.Index"/> is 0) this should be empty.
+        /// Goes to <see cref="PreviousHash"/>.</param>
+        /// <param name="txHash">The result of hashing the transactions the block has.
+        /// Goes to <see cref="TxHash"/>.</param>
+        /// <param name="preEvaluationHash">The hash derived from the block <em>excluding</em>
+        /// <paramref name="stateRootHash"/> (i.e., without action evaluation).
+        /// Used for checking <paramref name="nonce"/>.  See also <see cref="Validate"/>.</param>
+        /// <param name="stateRootHash">The <see cref="ITrie.Hash"/> of the states on the block.
+        /// </param>
+        internal BlockHeader(
+            int protocolVersion,
+            long index,
+            string timestamp,
+            ImmutableArray<byte> nonce,
+            ImmutableArray<byte> miner,
+            long difficulty,
+            BigInteger totalDifficulty,
+            ImmutableArray<byte> previousHash,
+            ImmutableArray<byte> txHash,
+            ImmutableArray<byte> preEvaluationHash,
+            ImmutableArray<byte> stateRootHash)
+        {
+            // FIXME: Basic sanity check, such as whether stateRootHash is empty or not,
+            // to prevent improper usage should be present. For the same reason as
+            // a comment in Block<T>(), should be added in on furter refactoring.
+            ProtocolVersion = protocolVersion;
+            Index = index;
+            Timestamp = timestamp;
+            Nonce = nonce;
+            Miner = miner;
+            Difficulty = difficulty;
+            TotalDifficulty = totalDifficulty;
+            PreviousHash = previousHash;
+            TxHash = txHash;
+
+            PreEvaluationHash = preEvaluationHash;
+            StateRootHash = stateRootHash;
+            Hash = BlockHash.DeriveFrom(SerializeForHash()).ByteArray;
         }
 
         /// <summary>
@@ -242,52 +358,6 @@ namespace Libplanet.Blocks
             return dict;
         }
 
-        internal static byte[] SerializeForHash(
-            int protocolVersion,
-            long index,
-            string timestamp,
-            long difficulty,
-            ImmutableArray<byte> nonce,
-            ImmutableArray<byte> miner,
-            ImmutableArray<byte> previousHash,
-            ImmutableArray<byte> txHash,
-            ImmutableArray<byte> stateRootHash
-        )
-        {
-            var dict = Bencodex.Types.Dictionary.Empty
-                .Add("index", index)
-                .Add("timestamp", timestamp)
-                .Add("difficulty", difficulty)
-                .Add("nonce", nonce.ToArray());
-
-            if (protocolVersion != 0)
-            {
-                dict = dict.Add("protocol_version", protocolVersion);
-            }
-
-            if (!miner.IsEmpty)
-            {
-                dict = dict.Add("reward_beneficiary", miner.ToArray());
-            }
-
-            if (!previousHash.IsEmpty)
-            {
-                dict = dict.Add("previous_hash", previousHash.ToArray());
-            }
-
-            if (!txHash.IsEmpty)
-            {
-                dict = dict.Add("transaction_fingerprint", txHash.ToArray());
-            }
-
-            if (!stateRootHash.IsEmpty)
-            {
-                dict = dict.Add("state_root_hash", stateRootHash.ToArray());
-            }
-
-            return new Codec().Encode(dict);
-        }
-
         internal void Validate(HashAlgorithmType hashAlgorithm, DateTimeOffset currentTime)
         {
             if (ProtocolVersion < 0)
@@ -305,13 +375,11 @@ namespace Libplanet.Blocks
                 throw new InvalidBlockProtocolVersionException(ProtocolVersion, message);
             }
 
+            BlockHash hash = new BlockHash(Hash);
             DateTimeOffset ts = DateTimeOffset.ParseExact(
                 Timestamp,
                 TimestampFormat,
-                CultureInfo.InvariantCulture
-            );
-
-            BlockHash hash = new BlockHash(Hash);
+                CultureInfo.InvariantCulture);
 
             if (currentTime + TimestampThreshold < ts)
             {
@@ -392,30 +460,82 @@ namespace Libplanet.Blocks
             {
                 throw new InvalidBlockNonceException(
                     $"Block #{Index} {hash}'s pre-evaluation hash " +
-                    $"({ByteUtil.Hex(PreEvaluationHash)}) with the nonce " +
+                    $"({ByteUtil.Hex(PreEvaluationHash)}) with nonce " +
                     $"({ByteUtil.Hex(Nonce)}) does not satisfy its difficulty level {Difficulty}."
                 );
             }
 
-            BlockHash calculatedHash = BlockHash.DeriveFrom(SerializeForHash());
-            if (!hash.Equals(calculatedHash))
+            // PreEvaluationHash comparison between the actual and the expected was not
+            // implemented in ProtocolVersion == 0.
+            if (ProtocolVersion > 0)
+            {
+                byte[] expectedPreEvaluationHash =
+                    hashAlgorithm.Digest(SerializeForPreEvaluationHash());
+                if (!ByteUtil.TimingSafelyCompare(expectedPreEvaluationHash, PreEvaluationHash))
+                {
+                    string message =
+                        $"The expected pre-evaluation hash of block #{Index} " +
+                        $"{ByteUtil.Hex(Hash)} is {ByteUtil.Hex(expectedPreEvaluationHash)}, " +
+                        $"but its pre-evaluation hash is {ByteUtil.Hex(PreEvaluationHash)}.";
+                    throw new InvalidBlockPreEvaluationHashException(
+                        PreEvaluationHash,
+                        expectedPreEvaluationHash.ToImmutableArray(),
+                        message);
+                }
+            }
+
+            BlockHash expectedHash = BlockHash.DeriveFrom(SerializeForHash());
+            if (!hash.Equals(expectedHash))
             {
                 throw new InvalidBlockHashException(
-                    $"The block #{Index} {hash}'s isn't matched its content, " +
-                    $"caculcated: {calculatedHash}");
+                    $"The expected hash {expectedHash} of block #{Index} {hash} does not match " +
+                    "the hash provided by the block.");
             }
         }
 
-        internal byte[] SerializeForHash(bool includeStateRootHash = true) => SerializeForHash(
-            ProtocolVersion,
-            Index,
-            Timestamp,
-            Difficulty,
-            Nonce,
-            Miner,
-            PreviousHash,
-            TxHash,
-            includeStateRootHash ? StateRootHash : ImmutableArray<byte>.Empty
-        );
+        internal Bencodex.Types.Dictionary ToBencodexForPreEvaluationHash()
+        {
+            var dict = Bencodex.Types.Dictionary.Empty
+                .Add("index", Index)
+                .Add("timestamp", Timestamp)
+                .Add("difficulty", Difficulty)
+                .Add("nonce", Nonce.ToArray())
+                .Add("reward_beneficiary", Miner.ToArray());
+
+            if (ProtocolVersion != 0)
+            {
+                dict = dict.Add("protocol_version", ProtocolVersion);
+            }
+
+            if (!PreviousHash.IsEmpty)
+            {
+                dict = dict.Add("previous_hash", PreviousHash.ToArray());
+            }
+
+            if (!TxHash.IsEmpty)
+            {
+                dict = dict.Add("transaction_fingerprint", TxHash.ToArray());
+            }
+
+            return dict;
+        }
+
+        internal Bencodex.Types.Dictionary ToBencodexForHash()
+        {
+            var dict = ToBencodexForPreEvaluationHash();
+
+            if (!StateRootHash.IsEmpty)
+            {
+                dict = dict.Add("state_root_hash", StateRootHash.ToArray());
+            }
+
+            return dict;
+        }
+
+        internal byte[] SerializeForPreEvaluationHash()
+            => new Codec().Encode(ToBencodexForPreEvaluationHash());
+
+        internal byte[] SerializeForHash()
+            => new Codec().Encode(ToBencodexForHash());
     }
 }